Separate display of synced/unsynced patterns
