+++ packages/edit-site/src/components/library/no-patterns.js
@@ -1,0 +1,12 @@
+/**
+ * WordPress dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+export default function NoPatterns() {
+	return (
+		<div className="edit-site-library__no-results">
+			{ __( 'No patterns found.' ) }
+		</div>
+	);
+}
