+++ packages/edit-site/src/components/library/use-patterns.js
@@ -16,6 +16,10 @@ const EMPTY_PATTERN_LIST = [];
 export const TEMPLATE_PARTS = 'wp_template_part';
 export const PATTERNS = 'pattern';
 export const USER_PATTERNS = 'wp_block';
+const SYNC_TYPES = {
+	full: 'fully',
+	unsynced: 'unsynced',
+};
 
 const createTemplatePartId = ( theme, slug ) =>
 	theme && slug ? theme + '//' + slug : null;
@@ -133,53 +137,81 @@ const reusableBlockToPattern = ( reusableBlock ) => ( {
 
 const useUserPatterns = ( categoryId, categoryType = PATTERNS ) => {
 	const postType = categoryType === PATTERNS ? USER_PATTERNS : categoryType;
-	const userPatterns = useSelect(
+	const unfilteredPatterns = useSelect(
 		( select ) => {
 			if ( postType !== USER_PATTERNS ) {
-				return EMPTY_PATTERN_LIST;
+				return {};
 			}
 
 			const { getEntityRecords } = select( coreStore );
-			const nonSyncedPatterns = getEntityRecords( 'postType', postType, {
+			const records = getEntityRecords( 'postType', postType, {
 				per_page: -1,
 			} );
 
-			if ( ! nonSyncedPatterns ) {
-				return EMPTY_PATTERN_LIST;
+			if ( ! records ) {
+				return {};
 			}
 
-			return nonSyncedPatterns.map( ( item ) =>
-				reusableBlockToPattern( item )
-			);
+			const patterns = { synced: [], unsynced: [] };
+
+			records.forEach( ( record ) => {
+				const pattern = reusableBlockToPattern( record );
+
+				if ( record.meta?.wp_block?.sync_status === SYNC_TYPES.full ) {
+					patterns.synced.push( pattern );
+				} else {
+					patterns.unsynced.push( pattern );
+				}
+			} );
+
+			return patterns;
 		},
 		[ postType ]
 	);
 
 	if ( categoryId === 'uncategorized' ) {
-		return userPatterns.filter(
-			( pattern ) => ! pattern.categories?.length
-		);
+		return {
+			syncedPatterns: unfilteredPatterns.synced?.filter(
+				( pattern ) => ! pattern.categories?.length
+			),
+			unsyncedPatterns: unfilteredPatterns.unsynced?.filter(
+				( pattern ) => ! pattern.categories?.length
+			),
+		};
 	}
 
 	const currentId = parseInt( categoryId );
 
-	return userPatterns.filter( ( pattern ) =>
-		pattern.categories?.includes( currentId )
-	);
+	return {
+		syncedPatterns: unfilteredPatterns.synced?.filter( ( pattern ) =>
+			pattern.categories?.includes( currentId )
+		),
+		unsyncedPatterns: unfilteredPatterns.unsynced?.filter( ( pattern ) =>
+			pattern.categories?.includes( currentId )
+		),
+	};
 };
 
 export const usePatterns = ( categoryType, categoryId ) => {
-	const patterns = useBlockPatternsByCategory( categoryId, categoryType );
-	const userPatterns = useUserPatterns( categoryId, categoryType );
+	const blockPatterns = useBlockPatternsByCategory(
+		categoryId,
+		categoryType
+	);
+	const { syncedPatterns = [], unsyncedPatterns = [] } = useUserPatterns(
+		categoryId,
+		categoryType
+	);
 	const { templateParts, isResolving } = useTemplatePartsAsPatterns(
 		categoryId,
 		categoryType
 	);
 
-	// TODO: Add sorting etc.
+	const patterns = {
+		syncedPatterns: [ ...templateParts, ...syncedPatterns ],
+		unsyncedPatterns: [ ...blockPatterns, ...unsyncedPatterns ],
+	};
 
-	const results = [ ...templateParts, ...patterns, ...userPatterns ];
-	return [ results, isResolving ];
+	return [ patterns, isResolving ];
 };
 
 export default usePatterns;
